requirements.txt
torch>=2.0.0
transformers>=4.35.0
rdkit-pypi>=2022.9.5
biopython>=1.83
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
tqdm>=4.65.0
